+++ mobile/android/android-components/components/feature/top-sites/src/main/java/mozilla/components/feature/top/sites/DefaultTopSitesStorage.kt
@@ -66,10 +66,10 @@ class DefaultTopSitesStorage(
         }
     }
 
-    override fun renameTopSite(topSite: TopSite, title: String) {
+    override fun updateTopSite(topSite: TopSite, title: String, url: String) {
         scope.launch {
             if (topSite.type != FRECENT) {
-                pinnedSitesStorage.renamePinnedSite(topSite, title)
+                pinnedSitesStorage.updatePinnedSite(topSite, title, url)
             }
 
             notifyObservers { onStorageUpdated() }